draw-cards-button: make misclicks harder
If user clicks Draw Cards within 6 seconds of advancing to a new turn for the
first time, we prompt the user if they're sure they're ready to draw cards.
The prompt is not shown if user has already been on the next turn, or if
more than 6 seconds have passed since the last action was recorded.